requirements/dev.txt
-r base.txt
django-debug-toolbar

behave
behave-django
factory_boy
django-nose
splinter
python-levenshtein
ipdb
